// models/ShopUser.ts
import { Title } from '../pages/automationExercise/AccountCreationPage';

export interface ShopUser {
  title: Title;
  password: string;
  birthDay: string;
  birthMonth: string;
  birthYear: string;
  name: string;
  email: string;
  company: string;
  firstName: string;
  lastName: string;
  address: string;
  address2: string;
  state: string;
  country: string;
  city: string;
  zipCode: string;
  mobileNumber: string;
}